Api Security: /api/vacations with OFFICE role and own data
